vfs: validate path components
Instead of making every fs implementation validating the path components
in API handlers, we can validate them in the VFS api handler.

Passthroughfs still keeps its own validation since it might be used
without a VFS layer.
