+++ src/api/mod.rs
@@ -17,7 +17,10 @@ pub use super::abi::linux_abi::CreateIn;
 mod pseudo_fs;
 
 pub mod vfs;
-pub use vfs::{BackendFileSystem, Vfs, VfsOptions, VFS_MAX_INO};
+pub use vfs::{
+    validate_path_component, BackFileSystem, BackendFileSystem, Vfs, VfsIndex, VfsOptions,
+    CURRENT_DIR_CSTR, EMPTY_CSTR, PARENT_DIR_CSTR, PROC_CSTR, SLASH_ASCII, VFS_MAX_INO,
+};
 
 pub mod filesystem;
 pub mod server;
