+++ src/api/vfs/sync_io.rs
@@ -64,6 +64,11 @@ impl<D: AsyncDrive, S: BitmapSlice> FileSystem<S> for Vfs<D, S> {
     }
 
     fn lookup(&self, ctx: &Context, parent: VfsInode, name: &CStr) -> Result<Entry> {
+        // Don't use is_safe_path_component(), allow "." and ".." for NFS export support
+        if name.to_bytes_with_nul().contains(&SLASH_ASCII) {
+            return Err(io::Error::from_raw_os_error(libc::EINVAL));
+        }
+
         match self.get_real_rootfs(parent)? {
             (Left(fs), idata) => self.lookup_pseudo(fs, idata, ctx, name),
             (Right(fs), idata) => {
@@ -128,6 +133,8 @@ impl<D: AsyncDrive, S: BitmapSlice> FileSystem<S> for Vfs<D, S> {
         parent: VfsInode,
         name: &CStr,
     ) -> Result<Entry> {
+        validate_path_component(name)?;
+
         match self.get_real_rootfs(parent)? {
             (Left(fs), idata) => fs.symlink(ctx, linkname, idata.ino(), name),
             (Right(fs), idata) => fs.symlink(ctx, linkname, idata.ino(), name).map(|mut e| {
@@ -146,6 +153,8 @@ impl<D: AsyncDrive, S: BitmapSlice> FileSystem<S> for Vfs<D, S> {
         rdev: u32,
         umask: u32,
     ) -> Result<Entry> {
+        validate_path_component(name)?;
+
         match self.get_real_rootfs(inode)? {
             (Left(fs), idata) => fs.mknod(ctx, idata.ino(), name, mode, rdev, umask),
             (Right(fs), idata) => {
@@ -166,6 +175,8 @@ impl<D: AsyncDrive, S: BitmapSlice> FileSystem<S> for Vfs<D, S> {
         mode: u32,
         umask: u32,
     ) -> Result<Entry> {
+        validate_path_component(name)?;
+
         match self.get_real_rootfs(parent)? {
             (Left(fs), idata) => fs.mkdir(ctx, idata.ino(), name, mode, umask),
             (Right(fs), idata) => fs.mkdir(ctx, idata.ino(), name, mode, umask).map(|mut e| {
@@ -176,6 +187,8 @@ impl<D: AsyncDrive, S: BitmapSlice> FileSystem<S> for Vfs<D, S> {
     }
 
     fn unlink(&self, ctx: &Context, parent: VfsInode, name: &CStr) -> Result<()> {
+        validate_path_component(name)?;
+
         match self.get_real_rootfs(parent)? {
             (Left(fs), idata) => fs.unlink(ctx, idata.ino(), name),
             (Right(fs), idata) => fs.unlink(ctx, idata.ino(), name),
@@ -183,6 +196,8 @@ impl<D: AsyncDrive, S: BitmapSlice> FileSystem<S> for Vfs<D, S> {
     }
 
     fn rmdir(&self, ctx: &Context, parent: VfsInode, name: &CStr) -> Result<()> {
+        validate_path_component(name)?;
+
         match self.get_real_rootfs(parent)? {
             (Left(fs), idata) => fs.rmdir(ctx, idata.ino(), name),
             (Right(fs), idata) => fs.rmdir(ctx, idata.ino(), name),
@@ -198,6 +213,9 @@ impl<D: AsyncDrive, S: BitmapSlice> FileSystem<S> for Vfs<D, S> {
         newname: &CStr,
         flags: u32,
     ) -> Result<()> {
+        validate_path_component(oldname)?;
+        validate_path_component(newname)?;
+
         let (root, idata_old) = self.get_real_rootfs(olddir)?;
         let (_, idata_new) = self.get_real_rootfs(newdir)?;
 
@@ -232,6 +250,8 @@ impl<D: AsyncDrive, S: BitmapSlice> FileSystem<S> for Vfs<D, S> {
         newparent: VfsInode,
         newname: &CStr,
     ) -> Result<Entry> {
+        validate_path_component(newname)?;
+
         let (root, idata_old) = self.get_real_rootfs(inode)?;
         let (_, idata_new) = self.get_real_rootfs(newparent)?;
 
@@ -276,6 +296,8 @@ impl<D: AsyncDrive, S: BitmapSlice> FileSystem<S> for Vfs<D, S> {
         name: &CStr,
         args: CreateIn,
     ) -> Result<(Entry, Option<u64>, OpenOptions)> {
+        validate_path_component(name)?;
+
         match self.get_real_rootfs(parent)? {
             (Left(fs), idata) => fs.create(ctx, idata.ino(), name, args),
             (Right(fs), idata) => {
@@ -426,6 +448,8 @@ impl<D: AsyncDrive, S: BitmapSlice> FileSystem<S> for Vfs<D, S> {
         value: &[u8],
         flags: u32,
     ) -> Result<()> {
+        validate_path_component(name)?;
+
         match self.get_real_rootfs(inode)? {
             (Left(fs), idata) => fs.setxattr(ctx, idata.ino(), name, value, flags),
             (Right(fs), idata) => fs.setxattr(ctx, idata.ino(), name, value, flags),
@@ -439,6 +463,8 @@ impl<D: AsyncDrive, S: BitmapSlice> FileSystem<S> for Vfs<D, S> {
         name: &CStr,
         size: u32,
     ) -> Result<GetxattrReply> {
+        validate_path_component(name)?;
+
         match self.get_real_rootfs(inode)? {
             (Left(fs), idata) => fs.getxattr(ctx, idata.ino(), name, size),
             (Right(fs), idata) => fs.getxattr(ctx, idata.ino(), name, size),
@@ -453,6 +479,8 @@ impl<D: AsyncDrive, S: BitmapSlice> FileSystem<S> for Vfs<D, S> {
     }
 
     fn removexattr(&self, ctx: &Context, inode: VfsInode, name: &CStr) -> Result<()> {
+        validate_path_component(name)?;
+
         match self.get_real_rootfs(inode)? {
             (Left(fs), idata) => fs.removexattr(ctx, idata.ino(), name),
             (Right(fs), idata) => fs.removexattr(ctx, idata.ino(), name),
