+++ src/passthrough/mod.rs
@@ -29,7 +29,10 @@ use vm_memory::ByteValued;
 
 use crate::abi::linux_abi as fuse;
 use crate::api::filesystem::Entry;
-use crate::api::{BackendFileSystem, VFS_MAX_INO};
+use crate::api::{
+    validate_path_component, BackendFileSystem, CURRENT_DIR_CSTR, EMPTY_CSTR, PARENT_DIR_CSTR,
+    PROC_CSTR, SLASH_ASCII, VFS_MAX_INO,
+};
 use crate::BitmapSlice;
 
 #[cfg(feature = "async-io")]
@@ -42,12 +45,6 @@ use file_handle::{FileHandle, MountFds};
 use multikey::MultikeyBTreeMap;
 
 use crate::async_util::{AsyncDrive, AsyncDriver};
-
-const CURRENT_DIR_CSTR: &[u8] = b".\0";
-const PARENT_DIR_CSTR: &[u8] = b"..\0";
-const EMPTY_CSTR: &[u8] = b"\0";
-const PROC_CSTR: &[u8] = b"/proc\0";
-const SLICE_ASCII: u8 = 47;
 
 type Inode = u64;
 type Handle = u64;
@@ -1024,30 +1021,6 @@ fn ebadf() -> io::Error {
     io::Error::from_raw_os_error(libc::EBADF)
 }
 
-#[inline]
-fn is_dot_or_dotdot(name: &CStr) -> bool {
-    let bytes = name.to_bytes_with_nul();
-    bytes.starts_with(CURRENT_DIR_CSTR) || bytes.starts_with(PARENT_DIR_CSTR)
-}
-
-// Is `path` a single path component that is not "." or ".."?
-fn is_safe_path_component(name: &CStr) -> bool {
-    let bytes = name.to_bytes_with_nul();
-
-    if bytes.contains(&SLICE_ASCII) {
-        return false;
-    }
-    !is_dot_or_dotdot(name)
-}
-
-#[inline]
-fn validate_path_component(name: &CStr) -> io::Result<()> {
-    match is_safe_path_component(name) {
-        true => Ok(()),
-        false => Err(io::Error::from_raw_os_error(libc::EINVAL)),
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -1185,69 +1158,6 @@ mod tests {
         assert_eq!(duration, fs.cfg.attr_timeout);
 
         fs.destroy();
-    }
-
-    #[test]
-    fn test_is_safe_path_component() {
-        let name = CStr::from_bytes_with_nul(b"normal\0").unwrap();
-        assert!(is_safe_path_component(name), "\"{:?}\"", name);
-
-        let name = CStr::from_bytes_with_nul(b".a\0").unwrap();
-        assert!(is_safe_path_component(name));
-
-        let name = CStr::from_bytes_with_nul(b"a.a\0").unwrap();
-        assert!(is_safe_path_component(name));
-
-        let name = CStr::from_bytes_with_nul(b"a.a\0").unwrap();
-        assert!(is_safe_path_component(name));
-
-        let name = CStr::from_bytes_with_nul(b"/\0").unwrap();
-        assert!(!is_safe_path_component(name));
-
-        let name = CStr::from_bytes_with_nul(b"/a\0").unwrap();
-        assert!(!is_safe_path_component(name));
-
-        let name = CStr::from_bytes_with_nul(b".\0").unwrap();
-        assert!(!is_safe_path_component(name));
-
-        let name = CStr::from_bytes_with_nul(b"..\0").unwrap();
-        assert!(!is_safe_path_component(name));
-
-        let name = CStr::from_bytes_with_nul(b"../.\0").unwrap();
-        assert!(!is_safe_path_component(name));
-
-        let name = CStr::from_bytes_with_nul(b"a/b\0").unwrap();
-        assert!(!is_safe_path_component(name));
-
-        let name = CStr::from_bytes_with_nul(b"./../a\0").unwrap();
-        assert!(!is_safe_path_component(name));
-    }
-
-    #[test]
-    fn test_is_dot_or_dotdot() {
-        let name = CStr::from_bytes_with_nul(b"..\0").unwrap();
-        assert!(is_dot_or_dotdot(name));
-
-        let name = CStr::from_bytes_with_nul(b".\0").unwrap();
-        assert!(is_dot_or_dotdot(name));
-
-        let name = CStr::from_bytes_with_nul(b"...\0").unwrap();
-        assert!(!is_dot_or_dotdot(name));
-
-        let name = CStr::from_bytes_with_nul(b"./.\0").unwrap();
-        assert!(!is_dot_or_dotdot(name));
-
-        let name = CStr::from_bytes_with_nul(b"a\0").unwrap();
-        assert!(!is_dot_or_dotdot(name));
-
-        let name = CStr::from_bytes_with_nul(b"aa\0").unwrap();
-        assert!(!is_dot_or_dotdot(name));
-
-        let name = CStr::from_bytes_with_nul(b"/a\0").unwrap();
-        assert!(!is_dot_or_dotdot(name));
-
-        let name = CStr::from_bytes_with_nul(b"a/\0").unwrap();
-        assert!(!is_dot_or_dotdot(name));
     }
 
     #[test]
