+++ src/passthrough/sync_io.rs
@@ -365,7 +365,7 @@ impl<D: AsyncDrive, S: BitmapSlice + Send + Sync> FileSystem<S> for PassthroughF
 
     fn lookup(&self, _ctx: &Context, parent: Inode, name: &CStr) -> io::Result<Entry> {
         // Don't use is_safe_path_component(), allow "." and ".." for NFS export support
-        if name.to_bytes_with_nul().contains(&SLICE_ASCII) {
+        if name.to_bytes_with_nul().contains(&SLASH_ASCII) {
             return Err(io::Error::from_raw_os_error(libc::EINVAL));
         }
         self.do_lookup(parent, name)
